Filter out running jobs in subquery
Jobs that are currently running have a time_start but time_end is 0.
With the old query, the metrics disappeared during the backup was
running.
With the new query the metric will keep returning the value from the
last **finished** backup job, which will make alerting easier and
produce graphs without gaps.